Dagger dependency injection: simplified declaration of cross-layer dependencies in the data/data-access layer (no more module 'initializer', Dagger is used to inject the component in the layer below the 'data')
