+++ sample-business-injection/src/main/java/com/teamwork/android/samples/business/injection/module/exported/DataRepoModule.java
@@ -1,6 +1,6 @@
 package com.teamwork.android.samples.business.injection.module.exported;
 
-import com.teamwork.android.samples.business.injection.module.bridge.SampleDataBridgeModule;
+import com.teamwork.android.samples.clean.data.access.DataAccessComponent;
 import com.teamwork.android.samples.clean.data.access.feature1.Entity1Repo;
 import com.teamwork.android.samples.clean.data.access.feature2.Entity2Repo;
 import dagger.Module;
@@ -10,16 +10,16 @@ import dagger.Provides;
  * Bridge Dagger module that exposes data repository dependencies from the data layer to the 'sample-business' module.
  */
 @Module
-public class DataRepoModule extends SampleDataBridgeModule {
+public class DataRepoModule {
 
     @Provides
-    Entity1Repo entity1Repo() {
-        return getComponent().entity1Repo();
+    Entity1Repo entity1Repo(DataAccessComponent component) {
+        return component.entity1Repo();
     }
 
     @Provides
-    Entity2Repo projectsRepo() {
-        return getComponent().entity2Repo();
+    Entity2Repo projectsRepo(DataAccessComponent component) {
+        return component.entity2Repo();
     }
 
 }
