+++ sample-data/src/main/java/com/teamwork/android/samples/clean/data/injection/SampleDataComponent.java
@@ -1,7 +1,8 @@
 package com.teamwork.android.samples.clean.data.injection;
 
-import com.teamwork.android.samples.clean.data.access.feature1.Entity1Repo;
-import com.teamwork.android.samples.clean.data.access.feature2.Entity2Repo;
+import android.support.annotation.NonNull;
+import android.support.annotation.RestrictTo;
+import com.teamwork.android.samples.clean.data.access.DataAccessComponent;
 import com.teamwork.android.samples.clean.data.injection.module.internal.DataRepoBindingModule;
 import com.teamwork.android.samples.clean.data.injection.module.internal.NetworkModule;
 import dagger.Component;
@@ -9,8 +10,11 @@ import dagger.Component;
 import javax.inject.Singleton;
 
 /**
- * Dagger component for dependency injection of classes in the 'sample-data' module and declaring which dependencies
- * need to be exposed to the business layer via a provision method (https://google.github.io/dagger/api/2.14/dagger/Component.html).
+ * Dagger component for dependency injection of classes in the 'sample-data' module (data layer).
+ * <p>
+ * The {@link Component} extends from {@link DataAccessComponent} so that the required "public" dependencies that need
+ * to be exposed to the business layer via a provision method (https://google.github.io/dagger/api/2.14/dagger/Component.html)
+ * are declared here.
  * <p>
  * TODO: provide an example of an AndroidModule (which provides the app Context and system services) that's used across
  * layers.
@@ -21,15 +25,42 @@ import javax.inject.Singleton;
         // dependency class bindings to expose modules
         DataRepoBindingModule.class
 })
-public interface SampleDataComponent {
+public interface SampleDataComponent extends DataAccessComponent {
 
-    /*
-     * Provision methods for public repositories that need to be exposed to the business layer.
-     * Any other repository implementation which is not declared here will only be accessible from the 'sample-data' module.
-     */
+    //region data layer internal dependencies
 
-    Entity1Repo entity1Repo();
+    /* Add here any dependencies which only need to be accessed (manually) from the data layer */
 
-    Entity2Repo entity2Repo();
+    //endregion
+
+    //region component provider
+
+    @SuppressWarnings("unused")
+    @RestrictTo(RestrictTo.Scope.LIBRARY)
+    static ComponentProvider getProvider() {
+        return ComponentProvider.COMPONENT_PROVIDER;
+    }
+
+    @SuppressWarnings("unused")
+    @RestrictTo(RestrictTo.Scope.LIBRARY)
+    class ComponentProvider {
+
+        private static final ComponentProvider COMPONENT_PROVIDER = new ComponentProvider();
+
+        private volatile SampleDataComponent dataComponent;
+
+        public void setComponent(@NonNull SampleDataComponent component) {
+            dataComponent = component;
+        }
+
+        public @NonNull SampleDataComponent getComponent() {
+            if (dataComponent == null) {
+                throw new IllegalStateException("SampleDataComponent not initialized");
+            }
+            return dataComponent;
+        }
+    }
+
+    //endregion
 
 }
